# docs/api/openapi.yml
components:
  schemas:
    AdoptIn:
      properties:
        berth_id:
          examples:
          - berth-001
          title: Berth Id
          type: string
        gateway_id:
          examples:
          - gw-dock-a
          title: Gateway Id
          type: string
        qr_payload:
          description: Base64url-encoded JSON from QR fragment (uuid, oob, sn, jwt)
          title: Qr Payload
          type: string
      required:
      - qr_payload
      - berth_id
      - gateway_id
      title: AdoptIn
      type: object
    AdoptionAdminOut:
      properties:
        berth_id:
          title: Berth Id
          type: string
        completed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Completed At
        created_at:
          format: date-time
          title: Created At
          type: string
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        error_msg:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Msg
        expires_at:
          format: date-time
          title: Expires At
          type: string
        gateway_id:
          title: Gateway Id
          type: string
        mesh_uuid:
          title: Mesh Uuid
          type: string
        request_id:
          title: Request Id
          type: string
        serial_number:
          title: Serial Number
          type: string
        status:
          examples:
          - pending
          title: Status
          type: string
      required:
      - request_id
      - mesh_uuid
      - serial_number
      - gateway_id
      - berth_id
      - status
      - expires_at
      - created_at
      title: AdoptionAdminOut
      type: object
    AdoptionCancelOut:
      properties:
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        request_id:
          title: Request Id
          type: string
        status:
          title: Status
          type: string
      required:
      - request_id
      - status
      title: AdoptionCancelOut
      type: object
...
    AdoptionRequestOut:
      properties:
        berth_id:
          examples:
          - berth-001
          title: Berth Id
          type: string
        completed_at:
          anyOf:
          - examples: &id001
            - '2026-05-03T14:31:00Z'
            format: date-time
            type: string
          - type: 'null'
          examples: *id001
          title: Completed At
        created_at:
          examples:
          - '2026-05-03T14:30:00Z'
          format: date-time
          title: Created At
          type: string
        error_code:
          anyOf:
          - examples: &id002
            - timeout
            type: string
          - type: 'null'
          examples: *id002
          title: Error Code
        error_msg:
          anyOf:
          - examples: &id003
            - node did not respond
            type: string
          - type: 'null'
          examples: *id003
          title: Error Msg
        expires_at:
          examples:
          - '2026-05-03T14:35:00Z'
          format: date-time
          title: Expires At
          type: string
        gateway_id:
          examples:
          - gw-dock-a
          title: Gateway Id
          type: string
        mesh_unicast_addr:
          anyOf:
          - examples: &id004
            - '0x0042'
            type: string
          - type: 'null'
          examples: *id004
          title: Mesh Unicast Addr
        mesh_uuid:
          examples:
          - a1b2c3d4-e5f6-7890-abcd-ef1234567890
          title: Mesh Uuid
          type: string
        request_id:
          examples:
          - req-0001
          title: Request Id
          type: string
        serial_number:
          examples:
          - DP-N-000123
          title: Serial Number
          type: string
        status:
          enum:
          - pending
          - ok
          - err
          title: Status
          type: string
      required:
      - request_id
      - mesh_uuid
      - serial_number
      - gateway_id
      - berth_id
      - status
      - expires_at
      - created_at
      title: AdoptionRequestOut
      type: object
...
    AssignmentOut:
      properties:
        berth_id:
          examples:
          - berth-001
          title: Berth Id
          type: string
        user_id:
          examples:
          - user-001
          title: User Id
          type: string
      required:
      - berth_id
      - user_id
      title: AssignmentOut
      type: object
    BerthAdminOut:
      properties:
        berth_id:
          title: Berth Id
          type: string
        depth_m:
          anyOf:
          - type: number
          - type: 'null'
          title: Depth M
        dock_id:
          title: Dock Id
          type: string
        is_reserved:
          title: Is Reserved
          type: boolean
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
        length_m:
          anyOf:
          - type: number
          - type: 'null'
          title: Length M
        status:
          examples:
          - free
          title: Status
          type: string
        width_m:
          anyOf:
          - type: number
          - type: 'null'
          title: Width M
      required:
      - berth_id
      - dock_id
      - status
      - is_reserved
      title: BerthAdminOut
      type: object
...
    BerthCreate:
      properties:
        berth_id:
          maxLength: 64
          minLength: 1
          title: Berth Id
          type: string
        depth_m:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - type: 'null'
          title: Depth M
        dock_id:
          maxLength: 64
          minLength: 1
          title: Dock Id
          type: string
        label:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Label
        length_m:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - type: 'null'
          title: Length M
        width_m:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - type: 'null'
          title: Width M
      required:
      - berth_id
      - dock_id
      title: BerthCreate
      type: object
    BerthCreatedOut:
      properties:
        berth_id:
          title: Berth Id
          type: string
        dock_id:
          title: Dock Id
          type: string
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
      required:
      - berth_id
      - dock_id
      title: BerthCreatedOut
      type: object
    BerthOut:
      properties:
        assignment:
          anyOf:
          - $ref: '#/components/schemas/AssignmentOut'
          - type: 'null'
        battery_pct:
          anyOf:
          - examples: &id005
            - 87
            type: integer
          - type: 'null'
          examples: *id005
          title: Battery Pct
        berth_id:
          examples:
          - berth-001
          title: Berth Id
          type: string
        depth_m:
          anyOf:
          - examples: &id006
            - 2.0
            type: number
          - type: 'null'
          examples: *id006
          title: Depth M
        dock_id:
          examples:
          - dock-a
          title: Dock Id
          type: string
        is_reserved:
          default: false
          title: Is Reserved
          type: boolean
        label:
          anyOf:
          - examples: &id007
            - A1
            type: string
          - type: 'null'
          examples: *id007
          title: Label
        last_updated:
          anyOf:
          - examples: &id008
            - '2026-05-03T14:30:00Z'
            format: date-time
            type: string
          - type: 'null'
          examples: *id008
          title: Last Updated
        length_m:
          anyOf:
          - examples: &id009
            - 8.5
            type: number
          - type: 'null'
          examples: *id009
          title: Length M
        sensor_raw:
          anyOf:
          - examples: &id010
            - 1234
            type: integer
          - type: 'null'
          examples: *id010
          title: Sensor Raw
        status:
          enum:
          - free
          - occupied
          title: Status
          type: string
        width_m:
          anyOf:
          - examples: &id011
            - 3.2
            type: number
          - type: 'null'
          examples: *id011
          title: Width M
      required:
      - berth_id
      - dock_id
      - status
      title: BerthOut
      type: object
    BerthPatch:
      properties:
        depth_m:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - type: 'null'
          title: Depth M
        is_reserved:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Reserved
        label:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Label
        length_m:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - type: 'null'
          title: Length M
        width_m:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - type: 'null'
          title: Width M
      title: BerthPatch
      type: object
    BerthPatchOut:
      properties:
        berth_id:
          title: Berth Id
          type: string
        depth_m:
          anyOf:
          - type: number
          - type: 'null'
          title: Depth M
        is_reserved:
          title: Is Reserved
          type: boolean
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
        length_m:
          anyOf:
          - type: number
          - type: 'null'
          title: Length M
        width_m:
          anyOf:
          - type: number
          - type: 'null'
          title: Width M
      required:
      - berth_id
      - is_reserved
      title: BerthPatchOut
      type: object
    BerthResetOut:
      properties:
        berth_id:
          title: Berth Id
          type: string
        status:
          title: Status
          type: string
      required:
      - berth_id
      - status
      title: BerthResetOut
      type: object
...
    BulkDeleteOut:
      properties:
        deleted:
          title: Deleted
          type: integer
        status_filter:
          title: Status Filter
          type: string
      required:
      - deleted
      - status_filter
      title: BulkDeleteOut
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    SweeperRunOut:
      properties:
        expired:
          title: Expired
          type: integer
        pruned:
          title: Pruned
          type: integer
      required:
      - expired
      - pruned
      title: SweeperRunOut
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    APIKeyCookie:
      in: cookie
      name: dockpulse_access
      type: apiKey
info:
  description: 'REST API for the DockPulse harbor berth availability system.


    The OpenAPI document at `/openapi.json` (and the YAML mirror in

    `docs/api/openapi.yml`) is generated directly from these route

    definitions and Pydantic models. The committed YAML is regenerated by

    `scripts/dump_openapi.py` and CI fails on drift.

    '
  title: DockPulse API
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/admin/adoptions:
    delete:
      description: default 'err', pass 'pending' to cancel-and-delete, 'all' to wipe
      operationId: adminBulkDeleteAdoptions
      parameters:
      - in: query
        name: status
        required: false
        schema:
          default: err
          title: Status
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDeleteOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Bulk Delete Adoptions
      tags:
      - admin
    get:
      operationId: adminListAdoptions
      parameters:
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - pattern: ^(pending|err|ok)$
            type: string
          - type: 'null'
          title: Status
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 500
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AdoptionAdminOut'
                title: Response Adminlistadoptions
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Adoptions
      tags:
      - admin
  /api/admin/adoptions/{request_id}/cancel:
    post:
      operationId: adminCancelAdoption
      parameters:
      - in: path
        name: request_id
        required: true
        schema:
          title: Request Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdoptionCancelOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Cancel Adoption
      tags:
      - admin
  /api/admin/berths:
    get:
      operationId: adminListBerths
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BerthAdminOut'
                title: Response Adminlistberths
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Berths
      tags:
      - admin
    post:
      operationId: adminCreateBerth
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BerthCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BerthCreatedOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Berth
      tags:
      - admin
  /api/admin/berths/{berth_id}:
    delete:
      operationId: adminDeleteBerth
      parameters:
      - in: path
        name: berth_id
        required: true
        schema:
          title: Berth Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Berth
      tags:
      - admin
    patch:
      operationId: adminPatchBerth
      parameters:
      - in: path
        name: berth_id
        required: true
        schema:
          title: Berth Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BerthPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BerthPatchOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Patch Berth
      tags:
      - admin
  /api/admin/berths/{berth_id}/reset:
    post:
      operationId: adminResetBerth
      parameters:
      - in: path
        name: berth_id
        required: true
        schema:
          title: Berth Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BerthResetOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reset Berth
      tags:
      - admin
...
  /api/admin/sweeper/run:
    post:
      operationId: adminRunSweeper
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SweeperRunOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Force a sweeper iteration (timeout expired pendings + prune old errs)
      tags:
      - admin
...
  /api/adoptions:
    post:
      operationId: createAdoption
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdoptIn'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdoptionRequestOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Create an adoption request for a scanned node
      tags:
      - adoptions
...
  /api/adoptions/{request_id}/cancel:
    post:
      operationId: cancelAdoption
      parameters:
      - in: path
        name: request_id
        required: true
        schema:
          title: Request Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdoptionRequestOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Cancel a pending adoption request
      tags:
      - adoptions
...
  /api/berths:
    get:
      operationId: listBerths
      parameters:
      - description: filter by dock
        in: query
        name: dock_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: filter by dock
          title: Dock Id
      - description: filter by status
        in: query
        name: status
        required: false
        schema:
          anyOf:
          - pattern: ^(free|occupied)$
            type: string
          - type: 'null'
          description: filter by status
          title: Status
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BerthOut'
                title: Response Listberths
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List all berths
      tags:
      - berths
...
  /api/berths/{berth_id}:
    get:
      operationId: getBerth
      parameters:
      - in: path
        name: berth_id
        required: true
        schema:
          title: Berth Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BerthOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a single berth
      tags:
      - berths
...
servers:
- description: Local development
  url: http://localhost:8000
- description: Staging
  url: https://staging.example.com
...